Finished Art
Blood Fractal- The symmetry of the recursive triangle yet slight bending of its lines represent the chaotic, organized chaos of nature. The red lines that web from its vertices allow you to create a 3d fractal using the base as your brush, symbolizing how all art stems from the chaos of the world we come from. If you click, the webbing will switch to randomized lines being drawn from the vertices, and the triangle stops wiggling, this is suppose to represent the opposite of the prior mode, un-chaotic base, yet chaotic brush, or art that stems from it. Yin and Yang so to speak. The webbing is red, as it represents the boldness of human nature/art. The background is white, as it represents life: or the canvas to be painted, a story to write.